+++ tests/test_env.py
@@ -1,5 +1,5 @@
 """
-    Copyright 2016 Inmanta
+    Copyright 2021 Inmanta
 
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
@@ -15,7 +15,7 @@
 
     Contact: bart@inmanta.com
 """
-
+import glob
 import logging
 import os
 import subprocess
@@ -73,13 +73,10 @@ def test_install_fails(tmpdir, caplog):
     venv.use_virtual_env()
     caplog.clear()
     package_name = "non-existing-pkg-inmanta"
-    exception_raised = False
-    try:
-        venv.install_from_list([package_name])
-    except Exception:
-        exception_raised = True
 
-    assert exception_raised
+    with pytest.raises(Exception):
+        venv.install_from_list([package_name])
+
     log_sequence = LogSequence(caplog)
     log_sequence.contains("inmanta.env", logging.ERROR, f"requirements: {package_name}")
 
@@ -99,7 +96,11 @@ def test_install_package_already_installed_in_parent_env(tmpdir):
     assert not set(parent_installed) - set(installed_packages)
 
     # site dir should be empty
-    site_dir = os.path.join(venv.env_path, "lib/python3.6/site-packages")
+    site_dir = os.path.join(venv.env_path, "lib/python*/site-packages")
+    dirs = glob.glob(site_dir)
+    assert len(dirs) == 1
+    site_dir = dirs[0]
+
     assert not os.listdir(site_dir)
 
     # test installing a package that is already present in the parent venv
@@ -151,41 +152,3 @@ def test_gen_req_file(tmpdir):
         'lorem == 0.1.1, > 0.1 ; python_version < "3.7" and platform_machine == "x86_64" and platform_system == "Linux"'
         in req_lines
     )
-
-
-def test_remove_requirements_present_in_parent_env(tmpdir):
-    v = env.VirtualEnv(tmpdir)
-    v.use_virtual_env()
-
-    # Verify parsing works
-    reqs = [
-        "lorem == 0.1.1",
-        "lorem > 0.1",
-        "lorem >= 1.1,<1.5",
-        "dummy-yummy",
-        "iplib@git+https://github.com/bartv/python3-iplib",
-    ]
-    assert v._remove_requirements_present_in_parent_env(reqs) == reqs
-
-    installed_packages = {
-        k: v for k, v in env.VirtualEnv._get_installed_packages(v._parent_python).items() if k != "pip" and k != "setuptools"
-    }
-    package = next(iter(installed_packages.keys()))
-    package_version = installed_packages[package]
-    splitted_package_version = package_version.split(".", maxsplit=1)
-    newer_package_version = f"{int(splitted_package_version[0]) + 1}.{splitted_package_version[1]}"
-
-    # Package is present in parent venv and constraint is met
-    reqs = [f"{package}=={package_version}", "non_existing_package==1.1.1"]
-    reqs_after_removal = ["non_existing_package==1.1.1"]
-    assert v._remove_requirements_present_in_parent_env(reqs) == reqs_after_removal
-
-    reqs = [f"{package}>={package_version},<{newer_package_version}", "non_existing_package==1.1.1"]
-    assert v._remove_requirements_present_in_parent_env(reqs) == reqs_after_removal
-
-    # Package is present in parent venv, but constraint isn't met
-    reqs = [f"{package}=={newer_package_version}", "non_existing_package==1.1.1"]
-    assert v._remove_requirements_present_in_parent_env(reqs) == reqs
-
-    reqs = [f"{package}>{package_version}", "non_existing_package==1.1.1"]
-    assert v._remove_requirements_present_in_parent_env(reqs) == reqs